Quality of Life changes
Changed the used card designs in Third Game, to make them more accessible for colour blind users, also added moving background for aesthetics.

Fixed issue in Fourth Game when the game wouldn't end after timer reached zero, which resulted in removing the collectible and not letting the player to do anything in particular.